// src/components/docs-sidebar-nav-item.tsx
import { component$ } from "@builder.io/qwik";
import { cn } from "~/lib/utils";
import { SidebarNavItem } from "~/types/nav"

interface DocsSidebarNavItemsProps {
  items: SidebarNavItem[]
  pathname: string | null
}

export default component$(({
  items,
  pathname,
}: DocsSidebarNavItemsProps) =>{
  return items?.length ? (
    <div class="grid grid-flow-row auto-rows-max text-sm">
      {items.map((item, index) =>
        item.href ? (
          <a
            key={index}
            href={item.href}
            class={cn(
              "group flex w-full items-center rounded-md border border-transparent px-2 py-1 hover:underline",
              item.disabled && "cursor-not-allowed opacity-60",
              pathname === item.href
                ? "font-medium text-foreground"
                : "text-muted-foreground"
            )}
            target={item.external ? "_blank" : ""}
            rel={item.external ? "noreferrer" : ""}
          >
            {item.title}
            {item.label && (
              <span class="ml-2 rounded-md bg-[#adfa1d] px-1.5 py-0.5 text-xs leading-none text-[#000000] no-underline group-hover:no-underline">
                {item.label}
              </span>
            )}
          </a>
        ) : (
          <span
            key={index}
            class="flex w-full cursor-not-allowed items-center rounded-md p-2 text-muted-foreground hover:underline"
          >
            {item.title}
          </span>
        )
      )}
    </div>
  ) : null
})